Fix NumberFormatException - use Long instead of Integer
There isn't any validation to check the value for 'memory' before converting it to an Integer so fixed that :)

At present, "34359738368234kB" and "34359738368234" throw a NumberFormatException since one is alphanumerical and the other exceeds the maximum value for the integer data type (both invalid int values). These are real values I retrieved from my API response.  

The calculation for formattedMemory fails unless both values above are handled so converting to a Long is probably the best way forward.

Morpheus forum post: 
- https://discuss.morpheusdata.com/t/datadog-plugin-is-calling-api-with-3f-instead-of-before-query-parameter/1764/14